added api requests
