--- app/src/main/java/com/example/searchmovieapp/contracts/HomeContract.kt
@@ -1,17 +1,0 @@
-package com.example.searchmovieapp.contracts
-
-import com.example.searchmovieapp.entities.MovieEntity
-
-interface HomeContract {
-    interface View {
-        fun showNowPlaying(nowPlayingMovies: List<MovieEntity>)
-        fun showUpcoming(upcomingMovies: List<MovieEntity>)
-    }
-
-    interface Presenter {
-        fun attach(view: View)
-        fun detach()
-        fun getMovies()
-        fun changeMovieFavoriteState(movieId: Int)
-    }
-}
