--- app/src/main/java/com/example/searchmovieapp/contracts/RatingsContract.kt
@@ -1,16 +1,0 @@
-package com.example.searchmovieapp.contracts
-
-import com.example.searchmovieapp.entities.MovieEntity
-
-interface RatingsContract {
-    interface View {
-        fun showMovies(movies: List<MovieEntity>)
-    }
-
-    interface Presenter {
-        fun attach(view: View)
-        fun detach()
-        fun getTopRatedMovies()
-        fun changeMovieFavoriteState(movieId: Int)
-    }
-}
