+++ app/src/main/java/com/example/searchmovieapp/data/RemoteDataSource.kt
@@ -4,10 +4,10 @@ import com.example.searchmovieapp.entities.MovieDetailsEntity
 import com.example.searchmovieapp.entities.MovieEntity
 
 interface RemoteDataSource {
-    suspend fun getNowPlayingMovies(): List<MovieEntity>
-    suspend fun getUpcomingMovies(): List<MovieEntity>
-    suspend fun getFavoritesMovies(): List<MovieEntity>
-    suspend fun getTopRatedMovies(): List<MovieEntity>
+    suspend fun getNowPlayingMovies(pageNum: Int): List<MovieEntity>
+    suspend fun getUpcomingMovies(pageNum: Int): List<MovieEntity>
+    suspend fun getFavoritesMovies(pageNum: Int): List<MovieEntity>
+    suspend fun getTopRatedMovies(pageNum: Int): List<MovieEntity>
     suspend fun getMovieDetailsById(movieId: Int): MovieDetailsEntity
     suspend fun setMovieAsFavorite(movieId: Int, isFavorite: Boolean)
 }
