+++ app/src/main/java/com/example/searchmovieapp/di/PresentersModule.kt
@@ -1,51 +1,43 @@
 package com.example.searchmovieapp.di
 
-import com.example.searchmovieapp.contracts.FavoritesContract
-import com.example.searchmovieapp.contracts.HomeContract
-import com.example.searchmovieapp.contracts.MovieDetailsContract
-import com.example.searchmovieapp.contracts.RatingsContract
-import com.example.searchmovieapp.presenters.FavoritesPresenter
-import com.example.searchmovieapp.presenters.HomePresenter
-import com.example.searchmovieapp.presenters.MovieDetailsPresenter
-import com.example.searchmovieapp.presenters.RatingsPresenter
 import com.example.searchmovieapp.repositories.MovieRepository
+import com.example.searchmovieapp.ui.details.MovieDetailsContract
+import com.example.searchmovieapp.ui.details.MovieDetailsPresenter
+import com.example.searchmovieapp.ui.favorites.FavoritesContract
+import com.example.searchmovieapp.ui.favorites.FavoritesPresenter
+import com.example.searchmovieapp.ui.home.HomeContract
+import com.example.searchmovieapp.ui.home.HomePresenter
+import com.example.searchmovieapp.ui.ratings.RatingsContract
+import com.example.searchmovieapp.ui.ratings.RatingsPresenter
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
-import dagger.hilt.android.components.FragmentComponent
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
 
-@InstallIn(FragmentComponent::class)
+@InstallIn(SingletonComponent::class)
 @Module
-class HomePresenterModule {
+class PresentersModule {
+    @Singleton
     @Provides
     fun provideHomePresenter(movieRepository: MovieRepository): HomeContract.Presenter {
         return HomePresenter(movieRepository)
     }
-}
 
-@InstallIn(FragmentComponent::class)
-@Module
-class MovieDetailsPresenterModule {
-    @Provides
-    fun provideMovieDetailsPresenter(movieRepository: MovieRepository): MovieDetailsContract.Presenter {
-        return MovieDetailsPresenter(movieRepository)
-    }
-}
-
-@InstallIn(FragmentComponent::class)
-@Module
-class FavoritesPresenterModule {
+    @Singleton
     @Provides
     fun provideFavoritesPresenter(movieRepository: MovieRepository): FavoritesContract.Presenter {
         return FavoritesPresenter(movieRepository)
     }
-}
 
-@InstallIn(FragmentComponent::class)
-@Module
-class RatingsPresenterModule {
+    @Singleton
     @Provides
     fun provideRatingsPresenter(movieRepository: MovieRepository): RatingsContract.Presenter {
         return RatingsPresenter(movieRepository)
+    }
+
+    @Provides
+    fun provideMovieDetailsPresenter(movieRepository: MovieRepository): MovieDetailsContract.Presenter {
+        return MovieDetailsPresenter(movieRepository)
     }
 }
