+++ app/src/main/java/com/example/searchmovieapp/entities/MovieEntity.kt
@@ -1,10 +1,18 @@
 package com.example.searchmovieapp.entities
 
+import com.squareup.moshi.Json
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
 data class MovieEntity(
     val id: Int,
     val title: String,
-    val posterPath: String?,
+    @Json(name = "poster_path")
+    var posterPath: String?,
+    @Json(name = "release_date")
     val releaseDate: String,
+    @Json(name = "vote_average")
     val voteAverage: Float,
-    val isUpcoming: Boolean
-)
+) {
+    var isUpcoming: Boolean = true
+}
