+++ app/src/main/java/com/example/searchmovieapp/repositories/MovieRepository.kt
@@ -15,22 +15,50 @@ private var favoriteMoviesIds: MutableList<Int> = mutableListOf()
 class MovieRepository(private val remoteDataSource: RemoteDataSource) {
     private val scope = CoroutineScope(Job() + Dispatchers.IO)
     private val job: Job = scope.launch {
-        favoriteMoviesIds = remoteDataSource.getFavoritesMovies().map { it.id }.toMutableList()
+        //todo implement getting all favorites movies
+        favoriteMoviesIds = remoteDataSource.getFavoritesMovies(1).map { it.id }.toMutableList()
     }
 
-    suspend fun getNowPlayingMovies(): List<MovieEntity> {
+    private val nowPlayingCache: MutableMap<Int, List<MovieEntity>> = mutableMapOf()
+    private val upcomingCache: MutableMap<Int, List<MovieEntity>> = mutableMapOf()
+    private val favoritesCache: MutableMap<Int, List<MovieEntity>> = mutableMapOf()
+    private val topRatedCache: MutableMap<Int, List<MovieEntity>> = mutableMapOf()
+
+    suspend fun getNowPlayingMovies(pageNum: Int): List<MovieEntity> {
         waitForFavoritesToLoad()
-        return remoteDataSource.getNowPlayingMovies()
+
+        if (!nowPlayingCache.containsKey(pageNum)) {
+            nowPlayingCache[pageNum] = remoteDataSource.getNowPlayingMovies(pageNum)
+        }
+
+        return nowPlayingCache.values.flatten()
     }
 
-    suspend fun getUpcomingMovies(): List<MovieEntity> {
+    suspend fun getUpcomingMovies(pageNum: Int): List<MovieEntity> {
         waitForFavoritesToLoad()
-        return remoteDataSource.getUpcomingMovies()
+
+        if (!upcomingCache.containsKey(pageNum)) {
+            upcomingCache[pageNum] = remoteDataSource.getUpcomingMovies(pageNum)
+        }
+
+        return upcomingCache.values.flatten()
     }
 
-    suspend fun getFavoritesMovies() = remoteDataSource.getFavoritesMovies()
+    suspend fun getFavoritesMovies(pageNum: Int): List<MovieEntity> {
+        if (!favoritesCache.containsKey(pageNum)) {
+            favoritesCache[pageNum] = remoteDataSource.getFavoritesMovies(pageNum)
+        }
 
-    suspend fun getTopRatedMovies() = remoteDataSource.getTopRatedMovies()
+        return favoritesCache.values.flatten()
+    }
+
+    suspend fun getTopRatedMovies(pageNum: Int): List<MovieEntity> {
+        if (!topRatedCache.containsKey(pageNum)) {
+            topRatedCache[pageNum] = remoteDataSource.getTopRatedMovies(pageNum)
+        }
+
+        return topRatedCache.values.flatten()
+    }
 
     suspend fun getMovieDetailsById(movieId: Int) = remoteDataSource.getMovieDetailsById(movieId)
