+++ app/src/main/java/com/example/searchmovieapp/ui/NavigationContainerFragment.kt
@@ -8,6 +8,11 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.example.searchmovieapp.R
 import com.example.searchmovieapp.databinding.FragmentNavigationContainerBinding
+import com.example.searchmovieapp.ui.favorites.FavoritesFragment
+import com.example.searchmovieapp.ui.home.HomeFragment
+import com.example.searchmovieapp.ui.ratings.RatingsFragment
+
+private const val CURRENT_FRAGMENT_TAG = "current frugment"
 
 class NavigationContainerFragment : Fragment() {
     private var _binding: FragmentNavigationContainerBinding? = null
@@ -48,12 +53,22 @@ class NavigationContainerFragment : Fragment() {
 
     private fun navigateToFragment(fragment: Fragment) {
         requireActivity().supportFragmentManager.beginTransaction()
-            .replace(R.id.container_with_navigation, fragment)
+            .replace(R.id.container_with_navigation, fragment, CURRENT_FRAGMENT_TAG)
             .commit()
     }
 
     override fun onDestroyView() {
-        super.onDestroyView()
         _binding = null
+
+        val currentFragment = requireActivity().supportFragmentManager
+            .findFragmentByTag(CURRENT_FRAGMENT_TAG)
+
+        currentFragment?.let {
+            requireActivity().supportFragmentManager.beginTransaction()
+                .remove(it)
+                .commit()
+        }
+
+        super.onDestroyView()
     }
 }
