+++ app/src/main/java/com/example/searchmovieapp/ui/details/MovieDetailsContract.kt
@@ -1,0 +1,16 @@
+package com.example.searchmovieapp.ui.details
+
+import com.example.searchmovieapp.entities.MovieDetailsEntity
+
+interface MovieDetailsContract {
+    interface View {
+        fun showDetails(movieDetails: MovieDetailsEntity)
+    }
+
+    interface Presenter {
+        fun attach(view: View)
+        fun detach()
+        fun getMovieDetails(movieId: Int)
+        fun changeMovieFavoriteState(movieId: Int)
+    }
+}
