+++ app/src/main/java/com/example/searchmovieapp/ui/favorites/FavoritesContract.kt
@@ -1,0 +1,22 @@
+package com.example.searchmovieapp.ui.favorites
+
+import android.os.Parcelable
+import com.example.searchmovieapp.entities.MovieEntity
+
+interface FavoritesContract {
+    interface View {
+        fun showFavorites(favoriteMovies: List<MovieEntity>)
+        fun restoreRecyclerViewPosition(position: Parcelable)
+    }
+
+    interface Presenter {
+        fun attach(view: View)
+        fun detach()
+        fun isFirstLoading(): Boolean
+        fun firstLoadingDone()
+        fun getMovies()
+        fun changeMovieFavoriteState(movieId: Int)
+        fun loadMore()
+        fun saveRecyclerViewPosition(position: Parcelable)
+    }
+}
