+++ app/src/main/res/layout/item_upcoming.xml
@@ -18,6 +18,7 @@
             android:id="@+id/poster_image_view"
             android:layout_width="0dp"
             android:layout_height="0dp"
+            android:scaleType="centerCrop"
             app:layout_constraintBottom_toTopOf="@id/title_text_view"
             app:layout_constraintDimensionRatio="1:1.5"
             app:layout_constraintEnd_toEndOf="parent"
