+++ app/src/main/res/values/strings.xml
@@ -11,4 +11,5 @@
     <string name="duration">%1$d min</string>
     <string name="release_date">Release date: %1$s</string>
     <string name="no_favorites">You have not mark any movie as favorite yet</string>
+    <string name="loading_error">Cannot load movies</string>
 </resources>
